requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
email_validator==2.0.0
requests==2.31.0
pyotp==2.8.0
qrcode==7.4.2
pillow==10.0.0
python-dotenv==1.0.0
